separate tests using claims format as JSON structure
